export: move deploy scripts away, add sup to kazoo container

+++ export/rabbitmq/run.sh
@@ -2,8 +2,14 @@
 FLAGS=${1:-"-td"}
 NETWORK=${NETWORK:-"kazoo"} 
 NAME=rabbitmq.$NETWORK
-docker stop -t 1 $NAME
-docker rm -f $NAME
+if [ -n "$(docker ps -aq -f name=$NAME)" ]
+then
+   echo -n "stopping: "
+   docker stop -t 1 $NAME
+   echo -n "removing: "
+   docker rm -f $NAME
+fi
+echo -n "starting: $NAME "
 docker run $FLAGS \
 	--net $NETWORK \
 	-h $NAME \
